% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning}
\alias{cleaning}
\title{Impute and Normalize a Data Frame}
\usage{
cleaning(
  data,
  zeros_as_na = FALSE,
  remove_zeros = FALSE,
  zero_cutoff = 0.3,
  na_cutoff = 0.3,
  knn = TRUE,
  norm = TRUE,
  first_col_ids = TRUE
)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Clean dataset.
}
\description{
This function imputes missing values and normalizes a data frame.
}
