% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_decimals.R
\name{detect_decimals}
\alias{detect_decimals}
\title{Detect Columns with Decimals in a Data Frame}
\usage{
detect_decimals(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A logical vector indicating which columns contain decimal values.
}
\description{
This function detects columns in a data frame that contain decimal values.
}
\examples{
data <- data.frame(x = c(1, 2, 3), y = c(1.5, 2.7, 3.0), z = c(1, 2, 3))
detect_decimals(data)

}
