% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_percentage.R
\name{normality_percentage}
\alias{normality_percentage}
\title{Normality in a Data Frame}
\usage{
normality_percentage(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Percentage of normailty.
}
\description{
This function computes the normality of all variables in a data frame.
}
